+++ src/widgets/index.ts
@@ -1,13 +1,14 @@
 // Экспорт всех виджетов
-export { default as ContactForm } from "./ContactForm";
-export { default as Footer } from "./Footer";
-export { default as Header } from "./Header";
-export { default as HeroSection } from "./HeroSection";
-export { default as ProductCard } from "./ProductCard";
-export { default as SiteBreadcrumbs } from "./SiteBreadcrumbs";
-export { default as ProductGallery } from "./ProductGallery";
-export { default as ProductHeader } from "./ProductHeader";
-export { default as ProductStatusChips } from "./ProductStatusChips";
-export { default as ProductPurchaseSection } from "./ProductPurchaseSection";
-export { default as ProductInfoBadges } from "./ProductInfoBadges";
-export { default as PromoConsultationCard } from "./PromoConsultationCard";
+export { default as ContactForm } from './ContactForm';
+export { default as Footer } from './Footer';
+export { default as Header } from './Header';
+export { default as HeroSection } from './HeroSection';
+export { default as ProductCard } from './ProductCard';
+export { ProductDescription } from './ProductDescription';
+export { default as ProductGallery } from './ProductGallery';
+export { default as ProductHeader } from './ProductHeader';
+export { default as ProductInfoBadges } from './ProductInfoBadges';
+export { default as ProductPurchaseSection } from './ProductPurchaseSection';
+export { default as ProductStatusChips } from './ProductStatusChips';
+export { default as PromoConsultationCard } from './PromoConsultationCard';
+export { default as SiteBreadcrumbs } from './SiteBreadcrumbs';